+++ libs/remix-ui/scriptrunner/src/types/index.ts
@@ -35,3 +35,11 @@ export interface ScriptRunnerConfig {
     customConfig: customScriptRunnerConfig
   }
 
+
+  export interface IScriptRunnerState {
+    customConfig: customScriptRunnerConfig
+    configurations: ProjectConfiguration[]
+    activeConfig: ProjectConfiguration
+    enableCustomScriptRunner: boolean
+  }
+
