+++ libs/remix-ui/tabs/src/lib/remix-ui-tabs.tsx
@@ -243,7 +243,7 @@ export const TabsUI = (props: TabsUIProps) => {
                 data-id="script-config"
                 className="btn text-dark border-left ml-2 pr-0 py-0 d-flex"
                 onClick={async () => {
-                  props.plugin.call('menuicons', 'select', 'scriptRunnerBridge')
+                  props.plugin.call('manager', 'activatePlugin', 'UIScriptRunner')
                 }}
               >
                 <i className="fa-kit fa-solid-gear-circle-play"></i>
